Charts: moved test utils to separate folder
