+++ packages/Charts/src/tests/radar-viewer-test.ts
@@ -3,7 +3,7 @@ import * as DG from 'datagrok-api/dg';
 
 import {after, before, category, expect, test} from '@datagrok-libraries/utils/src/test';
 import {Subscription} from 'rxjs';
-import {getOptions} from './utils';
+import {getOptions} from './utils/utils';
 
 
 category('Radar', () => {
