+++ packages/Charts/src/tests/timelines-viewer-test.ts
@@ -4,7 +4,7 @@ import * as DG from 'datagrok-api/dg';
 
 import {after, before, category, expect, test} from '@datagrok-libraries/utils/src/test';
 import {Subscription} from 'rxjs';
-import {getOptions} from './utils';
+import {getOptions} from './utils/utils';
 
 
 category('Timelines', () => {
