+++ packages/Charts/src/tests/tree-viewer-test.ts
@@ -2,7 +2,7 @@ import * as grok from 'datagrok-api/grok';
 import * as DG from 'datagrok-api/dg';
 
 import {after, before, category, expect, expectArray, test} from '@datagrok-libraries/utils/src/test';
-import {getOptions} from './utils';
+import {getOptions} from './utils/utils';
 
 
 category('TreeViewer', () => {
